+++ spring-cloud-starter-tencent-polaris-circuitbreaker/src/main/java/com/tencent/cloud/polaris/circuitbreaker/feign/PolarisFeignBeanPostProcessor.java
@@ -35,8 +35,7 @@ import org.springframework.cloud.openfeign.ribbon.LoadBalancerFeignClient;
  *
  * @author Haotian Zhang
  */
-public class PolarisFeignBeanPostProcessor
-		implements BeanPostProcessor, BeanFactoryAware {
+public class PolarisFeignBeanPostProcessor implements BeanPostProcessor, BeanFactoryAware {
 
 	private final ConsumerAPI consumerAPI;
 
@@ -47,8 +46,7 @@ public class PolarisFeignBeanPostProcessor
 	}
 
 	@Override
-	public Object postProcessBeforeInitialization(Object bean, String beanName)
-			throws BeansException {
+	public Object postProcessBeforeInitialization(Object bean, String beanName) throws BeansException {
 		return wrapper(bean);
 	}