+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - [Feature:Support near by router.](https://github.com/Tencent/spring-cloud-tencent/pull/196)
 - [Feature: Load application.yml and application-${profile}.yml from polaris server.](https://github.com/Tencent/spring-cloud-tencent/pull/199)
 - [feat:add rate limit of unirate.](https://github.com/Tencent/spring-cloud-tencent/pull/197)
+- [test:add junit test to polaris-circuitbreaker.](https://github.com/Tencent/spring-cloud-tencent/pull/202)